Replace the old dfx test with the new pocketic test for scenario-1
