+++ Cargo.toml
@@ -18,7 +18,6 @@ members = [
     "e2e-tests/scenario-3",
     "e2e-tests/cdk-bitcoin-canister",
     "e2e-tests/disable-api-if-not-fully-synced-flag",
-    "e2e-tests/scenario-1-pocketic",
     "test-utils",
     "ic-http",
     "ic-http/example_canister/src/canister_backend",
